+++ henpixy/tools/__init__.py
@@ -1,3 +1,5 @@
 """
-Módulo de ferramentas de processamento de imagens do Henpixy
-""" 
+Módulo com ferramentas para processamento de imagens
+"""
+from henpixy.tools.power import power_transform
+from henpixy.tools.contrast_stretching import contrast_stretching